requirements.txt
# requirements.txt
psycopg2-binary==2.9.9
sqlalchemy==2.0.28